Moodle export: better text input elements
In Moodle export we now specify for any "s" type question a wrong answer
"QQQQQQ" such that Moodle uses wider (and often equal length) text input
elements. (This way there are less or no hints about the length of the
correct answer. Also we had a report that chromium had so short elements
that answers with wide letters were not completely visible.